Make runtime directory optional
The runtime directory used to be used internally for a bunch of things, but at this point it's mainly used as a directory to write `edit` command buffers to.

Since those aren't crucial to the operation of PsySH, and since non-writable runtime directory configuration apparently happens sometimes, let's try making a writable runtime directory entirely optional.

See #443